requirements.txt
moviepy==1.0.3
pyautogui
Pillow
psutil
pygetwindow
opencv-python
send2trash